feat(creator): add static data tools
