+++ src/creator/api/getItems.ts
@@ -1,0 +1,8 @@
+import axios from 'axios';
+
+const getItems = ({ version, lang }) => {
+  const url = `https://ddragon.leagueoflegends.com/cdn/${version}/data/${lang}/item.json`;
+  return axios.get(url).then(response => response.data);
+};
+
+export default getItems;
