+++ src/creator/api/getSummonerSpells.ts
@@ -1,0 +1,8 @@
+import axios from 'axios';
+
+const getSummonerSpells = ({ version, lang }) => {
+  const url = `https://ddragon.leagueoflegends.com/cdn/${version}/data/${lang}/summoner.json`;
+  return axios.get(url).then(response => response.data);
+};
+
+export default getSummonerSpells;
