+++ src/creator/package.json
@@ -9,6 +9,7 @@
   "dependencies": {
     "@zeit/next-css": "^1.0.1",
     "@zeit/next-typescript": "^1.1.1",
+    "axios": "^0.18.0",
     "isomorphic-unfetch": "^3.0.0",
     "jsoneditor": "^5.27.0",
     "next": "^7.0.2",
